chore: responsive on prosessing 2
