+++ app/(general)/performances/[id]/teams/[teamId]/_components/BasicInfo.tsx
@@ -1,12 +1,6 @@
-import { useRouter } from "next/navigation"
-
-import { useToast } from "@/components/hooks/use-toast"
 import FreshmenFixedBadge from "@/components/TeamBadges/FreshmenFixedBadge"
 import StatusBadge from "@/components/TeamBadges/StatusBadge"
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import API_ENDPOINTS, { ApiEndpoint } from "@/constants/apiEndpoints"
-import ROUTES from "@/constants/routes"
-import fetchData from "@/lib/fetch"
 import {
   formatGenerationOrder,
   getRepresentativeRelativeTime
@@ -14,41 +8,11 @@ import {
 import { MemberSessionSet, Team } from "@/types/Team"
 
 interface BasicInfoProps {
-  performanceId: number
   team: Team
-  accessToken?: string
 }
 
-const BasicInfo = ({ performanceId, team, accessToken }: BasicInfoProps) => {
-  const router = useRouter()
-  const { toast } = useToast()
-
+const BasicInfo = ({ team }: BasicInfoProps) => {
   const memberSessionSet = new MemberSessionSet(team.memberSessions ?? [])
-
-  function onDeleteButtonClick() {
-    fetchData(API_ENDPOINTS.TEAM.DELETE(team.id) as ApiEndpoint, {
-      cache: "no-cache",
-      headers: {
-        Authorization: `Bearer ${accessToken}`
-      }
-    }).then(async (res) => {
-      if (!res.ok) {
-        switch (res.status) {
-          default:
-            toast({
-              title: "팀 삭제 실패",
-              description: "알 수 없는 이유",
-              variant: "destructive"
-            })
-        }
-      }
-      toast({
-        title: "팀 삭제 성공",
-        description: "성공적으로 팀을 삭제했습니다."
-      })
-      router.push(ROUTES.PERFORMANCE.TEAM.LIST(performanceId))
-    })
-  }
 
   return (
     <div className="w-full rounded-2xl bg-white px-12 pb-12 pt-14 shadow-md min-[600px]:px-20 min-[878px]:w-11/12 lg:w-5/6">
@@ -64,47 +28,31 @@ const BasicInfo = ({ performanceId, team, accessToken }: BasicInfoProps) => {
           {formatGenerationOrder(team.leader.generation.order)}기&nbsp;
           {team.leader.name}
         </div>
-        <div className="text-gray-400">
+        <div className="text-sm text-gray-400">
           {getRepresentativeRelativeTime(team.createdDatetime)}
         </div>
       </div>
       <div className="flex items-center justify-between">
         <div>
           {/* 곡 정보 */}
-          <div className="flex items-center gap-x-3">
-            <h3 className="text-3xl font-extrabold">{team.songName}</h3>
+          <div className="flex items-center gap-x-5 md:gap-x-3">
+            <h3 className="text-2xl font-extrabold md:text-3xl">
+              {team.songName}
+            </h3>
             <StatusBadge
               status={memberSessionSet.isSatisfied ? "Inactive" : "Active"}
             />
           </div>
-          <h4 className="my-2 text-xl text-gray-500">
+          <h4 className="my-2 font-['Inter'] text-lg font-normal text-gray-500 md:text-xl ">
             {team.songArtist} {team.isSelfMade && "(자작곡)"}
           </h4>
         </div>
-        {/*}
-        <div className="flex items-center justify-center gap-x-5">
-          <Button asChild variant="outline" className="h-12 w-12 p-2 shadow">
-            <Link
-              href={ROUTES.PERFORMANCE.TEAM.EDIT(team.performance.id, team.id)}
-            >
-              <PenLine strokeWidth={1.25} />
-            </Link>
-          </Button>
-          <form action={onDeleteButtonClick}>
-            <Button
-              type="submit"
-              variant="outline"
-              className="h-12 w-12 p-2 shadow"
-            >
-              <Trash2 strokeWidth={1.25} />
-            </Button>
-          </form>
-        </div>
-        */}
       </div>
 
       {/* 설명 */}
-      <p className="pb-4 pt-8">{team.description}</p>
+      <p className="pb-4 pt-8 font-['Inter'] text-base font-normal leading-normal tracking-wide text-slate-700 md:text-lg">
+        {team.description}
+      </p>
     </div>
   )
 }
